cloud-functions/pubsub-publisher/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.18.4
google-events==0.12.0
cloudevents==1.10.1
protobuf==4.25.1
